Refactoring of the model structure in furniture
- Renamed abstract classes: PlacementCoordinates → AbstractRoomPlacement, LayoutPlacement → AbstractLayoutPlacement.
- Created a new abstract_models.py file for abstract models.
- Split placements.py into two files: room_placements.py and layout_placements.py.
