+++ furniture/models/__init__.py
@@ -1,5 +1,5 @@
-from .base import PlacementCoordinates
-from .placements import (DoorPlacement, FurniturePlacement,
-                         PowerSocketPlacement, WindowPlacement)
+from .abstract_models import AbstractRoomPlacement
+from .layout_placements import FurniturePlacement, PowerSocketPlacement
 from .room_items import Furniture, RoomType
 from .room_layout import RoomLayout
+from .room_placements import DoorPlacement, WindowPlacement
